requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
sqlalchemy==2.0.40
psycopg[binary]==3.2.13
redis==5.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
jinja2==3.1.6
pydantic-settings==2.8.1
email-validator==2.2.0
alembic==1.15.2
bcrypt==4.0.1
